+++ CHANGELOG.md
@@ -11,6 +11,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Update default Affiliate Store Template
 
+## [0.14.0] - 2021-12-17
+
+## Added
+
+- Changed data entity being used to store Affiliate Lead from CL to a new entity.
+
 ## [0.13.0] - 2021-12-13
 
 ### Added